requirements.txt
pandas>=1.3.0
requests>=2.26.0
openpyxl>=3.0.7
python-dotenv>=1.0.0
